Centralize Fuel-defined exceptions.
